nexus/requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.29
httpx>=0.27
pydantic>=2.0
pyyaml>=6.0
psutil>=5.9
textblob>=0.18
gtts>=2.5
SpeechRecognition>=3.10
ib-insync>=0.9.86
python-dotenv>=1.0
websockets>=12.0
